+++ src/app/find/page.tsx
@@ -1,8 +1,7 @@
 "use client";
-import { useState, useEffect, useRef, useCallback } from "react";
+import { useState, useEffect, useRef, useCallback, Suspense } from "react";
 import { getHashtags } from "@/lib/fetchs/fetchHashtags";
 import { THashtag } from "@/lib/types/THashtag";
-import { TCocktail } from "@/lib/types/TCocktail";
 import { useOffsetStore } from "@/lib/store/offsetStore";
 import { useCocktailStore } from "@/lib/store/cocktailStore";
 import { getCocktail } from "@/lib/fetchs/fetchCocktail";
@@ -12,10 +11,12 @@ import style from "./Find.module.scss";
 import FindSearchBar from "@/components/find_search_bar/FindSearchBar";
 import CocktailList from "@/components/main/cocktailList/CocktailList";
 import Navigation from "@/components/common/navigation/Navigation";
+import SuspenseSearchParams from "./SuspenseSearchParams";
+
+export const dynamic = "force-dynamic";
 
 export default function FindPage() {
   const { cocktailList, totalCount } = useCocktailStore();
-  const [localCocktailList, setLocalCocktailList] = useState<TCocktail[]>([]);
   const { offset, setOffset } = useOffsetStore();
   const isLoading = useRef(false);
   const [inputValue, setInputValue] = useState("");
@@ -23,76 +24,66 @@ export default function FindPage() {
   const [hashtags, setHashtags] = useState<THashtag[]>([]);
   const [clickedHashtag, setClickedHashtag] = useState("");
 
-  const searchParams = useSearchParams();
+  const visibleCocktails = cocktailList.slice(0, offset);
 
   useEffect(() => {
     const fetchHashtags = async () => {
       const response = await getHashtags();
-      console.log("😎😋해쉬태그 가져왔다", response);
       setHashtags(response);
     };
     fetchHashtags();
   }, []);
   useEffect(() => {
-    const linkTop100 = searchParams.get("linkTop100");
-    if (linkTop100 === "1") {
-      setClickedHashtag("top100");
+    if (offset === 0) {
+      setOffset(25); // 초기 로딩
     }
-  }, [searchParams]);
+  }, []);
   //스크롤 바가 아래로 내려가면 실행될 loadMore 함수, 이 메서드는 cocktailList 컴포넌트로 넘겨준다.
   const loadMore = useCallback(async () => {
-    //isLoading.current가 true면 return
     if (isLoading.current) return;
     isLoading.current = true;
 
-    //칵테일이 모두 렌더링 된다면 isLoading.current를 false로 바꿔주어 loadMore가 실행되지 않도록 한다.
-    if (cocktailList.length === 0 || localCocktailList.length >= totalCount) {
+    if (cocktailList.length === 0 || offset >= totalCount) {
       isLoading.current = false;
       return;
     }
-    // 모든 칵테일 데이터가 들어왔다면 zustand 칵테일 데이터 기반으로 렌더링한다.
-    if (cocktailList.length != 0) {
-      setLocalCocktailList((prev) => {
-        return [...prev, ...cocktailList.slice(prev.length, prev.length + 25)];
-      });
-      setOffset(offset + 25);
-      console.log("이건 모든 칵테일 데이터로부터 옴😍😍😎😎😋");
-    } else {
-      //칵테일 데이터가 없을 때 db에서 25개씩 불러온다.
+
+    // 필요한 경우 getCocktail로 더 불러오기
+    if (cocktailList.length < offset + 25) {
       const newCocktails = await getCocktail(25, offset);
-      setLocalCocktailList((prev) => {
-        return [...prev, ...newCocktails.cocktails];
-      });
-      setOffset(offset + 25);
-      console.log("이건 25개 호출 데이터💥💥💥💢");
+      // 상태 갱신 로직은 기존 useCocktailStore 안에서 처리한다고 가정
     }
 
+    setOffset(offset + 25);
     isLoading.current = false;
-  }, [offset, localCocktailList.length, totalCount, setOffset, cocktailList]);
+  }, [offset, totalCount, cocktailList.length]);
 
   if (isLoading.current) {
     return <div className={`${style.loading}`}>Loading...</div>;
   }
 
   return (
-    <div className={`${style.body}`}>
-      <Navigation />
-      <FindSearchBar
-        hashtagList={hashtags}
-        onInputChange={setInputValue}
-        onSelectChange={setSelectValue}
-        onClickedHashtag={setClickedHashtag}
-        clickedHashtag={setClickedHashtag}
-        className={`${style.find_search_bar}`}
-      />
-      <CocktailList
-        cocktailList={localCocktailList}
-        loadMore={loadMore}
-        loading={isLoading.current}
-        inputValue={inputValue}
-        selectValue={selectValue}
-        clickedHashtag={clickedHashtag}
-      />
-    </div>
+    <Suspense fallback={<div>Loading...</div>}>
+      <div className={`${style.body}`}>
+        <Navigation />
+        <SuspenseSearchParams setClickedHashtag={setClickedHashtag} />
+        <FindSearchBar
+          hashtagList={hashtags}
+          onInputChange={setInputValue}
+          onSelectChange={setSelectValue}
+          onClickedHashtag={setClickedHashtag}
+          clickedHashtag={clickedHashtag}
+          className={`${style.find_search_bar}`}
+        />
+        <CocktailList
+          cocktailList={visibleCocktails}
+          loadMore={loadMore}
+          loading={isLoading.current}
+          inputValue={inputValue}
+          selectValue={selectValue}
+          clickedHashtag={clickedHashtag}
+        />
+      </div>
+    </Suspense>
   );
 }
