+++ src/app/storage/page.tsx
@@ -1,7 +1,7 @@
 import Storage from "./Storage";
 
 import { getServerSession } from "next-auth";
-import { authOptions } from "../api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/authOptions";
 
 export default async function StoragePage() {
   const session = await getServerSession(authOptions);
